Reduce duplication in indentation cops
Introduce `ArrayHashIndentation` to be used by `Style/IndentArray` and
`Style/IndentHash`. This should bring these up from the CodeClimate's
F-class.
