+++ lib/rubocop.rb
@@ -43,6 +43,7 @@ require 'rubocop/cop/variable_force/variable_table'
 
 require 'rubocop/cop/mixin/access_modifier_node'
 require 'rubocop/cop/mixin/annotation_comment'
+require 'rubocop/cop/mixin/array_hash_indentation'
 require 'rubocop/cop/mixin/array_syntax'
 require 'rubocop/cop/mixin/autocorrect_alignment'
 require 'rubocop/cop/mixin/autocorrect_unless_changing_ast'
